[test] gathering batched operations tests
